backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
websockets==12.0
python-socketio==5.11.1
openai-whisper==20231117
torch==2.2.1
numpy==1.26.4